+++ examples/static-spark-logs/emr.tf
@@ -2,26 +2,6 @@ locals {
   this_application = ["Spark"]
 }
 
-# Set up logs bucket with read/write permissions
-module "emr-logs-bucket" {
-  source      = "git::git@github.com:Datatamer/terraform-aws-s3.git?ref=1.1.0"
-  bucket_name = var.bucket_name_for_logs
-  read_write_actions = [
-    "s3:HeadBucket",
-    "s3:PutObject",
-  ]
-  read_write_paths = [""] # r/w policy permitting specified rw actions on entire bucket
-  tags             = var.tags
-}
-
-# Set up root directory bucket
-module "emr-rootdir-bucket" {
-  source           = "git::git@github.com:Datatamer/terraform-aws-s3.git?ref=1.1.0"
-  bucket_name      = var.bucket_name_for_root_directory
-  read_write_paths = [""] # r/w policy permitting default rw actions on entire bucket
-  tags             = var.tags
-}
-
 # Create new EC2 key pair
 resource "tls_private_key" "emr_private_key" {
   algorithm = "RSA"
@@ -47,17 +27,26 @@ module "emr-spark" {
   emr_config_file_path  = "../emr-config-template.json"
   tags                  = var.tags
   abac_valid_tags       = var.abac_valid_tags
+  bootstrap_actions = [
+    {
+      name = "cw_agent_install",
+      path = "s3://${module.emr-rootdir-bucket.bucket_name}/${aws_s3_bucket_object.sample_bootstrap_script.id}"
+      args = []
+    }
+  ]
 
   # Networking
-  subnet_id = var.subnet_id
+  subnet_id = var.compute_subnet_id
   vpc_id    = var.vpc_id
 
   # External resource references
   bucket_name_for_root_directory = module.emr-rootdir-bucket.bucket_name
   bucket_name_for_logs           = module.emr-logs-bucket.bucket_name
-  additional_policy_arns         = [module.emr-logs-bucket.rw_policy_arn, module.emr-rootdir-bucket.rw_policy_arn]
-  bucket_path_to_logs            = "logs/spark-test-cluster/"
-  key_pair_name                  = module.emr_key_pair.key_pair_key_name
+  additional_policy_arns = [module.emr-logs-bucket.rw_policy_arn,
+    module.emr-rootdir-bucket.rw_policy_arn,
+  "arn:aws:iam::aws:policy/CloudWatchAgentServerPolicy"]
+  bucket_path_to_logs = "logs/spark-test-cluster/"
+  key_pair_name       = module.emr_key_pair.key_pair_key_name
 
   # Names
   cluster_name                  = format("%s-%s", var.name_prefix, "Spark-Test-EMR-Cluster")
@@ -84,43 +73,4 @@ module "emr-spark" {
   emr_managed_master_sg_ids = module.aws-emr-sg-master.security_group_ids
   emr_managed_core_sg_ids   = module.aws-emr-sg-core.security_group_ids
   emr_service_access_sg_ids = module.aws-emr-sg-service-access.security_group_ids
-}
-
-module "sg-ports" {
-  # source  = "git::https://github.com/Datatamer/terraform-aws-emr.git//modules/aws-emr-ports?ref=7.3.0"
-  source       = "../../modules/aws-emr-ports"
-  applications = local.this_application
-}
-
-module "aws-emr-sg-master" {
-  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
-  vpc_id              = var.vpc_id
-  ingress_cidr_blocks = var.ingress_cidr_blocks
-  egress_cidr_blocks  = var.egress_cidr_blocks
-  ingress_ports       = module.sg-ports.ingress_master_ports
-  sg_name_prefix      = format("%s-%s", var.name_prefix, "-master")
-  egress_protocol     = "all"
-  ingress_protocol    = "tcp"
-}
-
-module "aws-emr-sg-core" {
-  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
-  vpc_id              = var.vpc_id
-  ingress_cidr_blocks = var.ingress_cidr_blocks
-  egress_cidr_blocks  = var.egress_cidr_blocks
-  ingress_ports       = module.sg-ports.ingress_core_ports
-  sg_name_prefix      = format("%s-%s", var.name_prefix, "-core")
-  egress_protocol     = "all"
-  ingress_protocol    = "tcp"
-}
-
-module "aws-emr-sg-service-access" {
-  source              = "git::git@github.com:Datatamer/terraform-aws-security-groups.git?ref=1.0.0"
-  vpc_id              = var.vpc_id
-  ingress_cidr_blocks = var.ingress_cidr_blocks
-  egress_cidr_blocks  = var.egress_cidr_blocks
-  ingress_ports       = module.sg-ports.ingress_service_access_ports
-  sg_name_prefix      = format("%s-%s", var.name_prefix, "-service-access")
-  egress_protocol     = "all"
-  ingress_protocol    = "tcp"
 }